// ai-app-doctor/src/types/tool.ts
export enum Tool {
    AnalyzeFile = "analyze_file",
    ReadFile = "read_file",
    ListFiles = "list_files",
    ListUntestedFiles = "list_untested_files",
    WriteFile = "write_file",
    WriteTestFile = "write_test_file",
    RunCommand = "run_command",
    RunTests = "run_tests",
    Git = "git",
    Gh = "gh",
    ScanProject = "scan_project",
}